Update uninstall script
Add env sourcing
Add context switch and checking
Add message to warning your about bucket

+++ vault-deploy/vault_server_setup/uninstall_vault_server.sh
@@ -11,6 +11,25 @@ then
     exit 0;
 fi
 
+printf "Sourcing envs from .env file\n";
+# Check if .env exists
+if [ -f ../.env ];
+then
+    source ../.env;
+else
+    printf "Error: .env file not found\n";
+    exit 1;
+fi
+
+printf "\nChanging context to ${CONTEXT_VAULT_CLUSTER}\n";
+kubectl config use-context $CONTEXT_VAULT_CLUSTER
+
+if [ $? != 0 ];
+then
+    printf "Aborting: Can't change context to %s\n" $CONTEXT_VAULT_CLUSTER;
+    exit 1;
+fi
+
 printf "\nUninstalling Vaul release (HELM)\n"
 helm uninstall vault -n vault
 
@@ -18,3 +37,5 @@ printf "\nRemoving remaining resources from cluster\n"
 kubectl delete secrets vault-server-tls -n vault
 kubectl delete csr vault-csr
 kubectl delete ns vault
+
+printf "\nDelete your bucket content manually\n"
